Added Linux port and bones of Zenoy architecture

+++ Core/rounding.cpp
@@ -2,11 +2,11 @@
 
 int TMath::round_nearest(double x)
 {
-	return floor(x + 0.5);
+	return (int) floor(x + 0.5);
 }
 
 int TMath::round_nearest(float f)
 {
 	float aftersum = f + 0.5f;
-	return floor(aftersum);
+	return (int) floor(aftersum);
 }
